app/Python/requirements.txt
PyMuPDF==1.23.14
pyzbar==0.1.9
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.4